requirements.txt
streamlit
numpy
plotly
pandas
requests
python-dotenv
scipy

